changed enroll class into user class
